+++ CHANGELOG.md
@@ -29,6 +29,7 @@ generating plots of HRRR ptype.
 - Accomodate ancient LSRs using `TRACE` as the magnitude field.
 - Ensure geometries going into masking helper are CCW, to mask outside of.
 - Improve dev experience using `ugc.UGCProvider` (#980).
+- Fix grid and affine definitions for IEMRE and MRMS.
 - Properly check USDM service response prior to parsing it.
 - Refine which `MWW` products are not defaulted into the main WFO Channels,
   those being VTEC codes `SC`, `MF`, and `GL`.